Type:fix bug 25139
Description:Failed to locate the schedule display closest to the date of the day in searchview

RootCause:Null

Solution:

+++ src/monthgraphiview.cpp
@@ -789,6 +789,8 @@ void CMonthGraphiview::slotDeleteItem()
 {
     if (m_infoitem == nullptr)
         return;
+    if (m_infoitem->getData().type.ID ==4)
+        return;
     DeleteItem(m_infoitem);
     m_infoitem =nullptr;
 }
